+++ app/Http/Controllers/ProfessoresController.php
@@ -17,6 +17,11 @@ class ProfessoresController extends Controller
      */
     public function index()
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_users) {
+            return abort(404, 'Unauthorized');
+        }
+
         $professores = User::teachers();
 
         return view('professor.index',['professores' => $professores]);
@@ -27,6 +32,10 @@ class ProfessoresController extends Controller
      */
     public function create()
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_users) {
+            return abort(404, 'Unauthorized');
+        }
         return view('professor.create');
     }
 
@@ -35,6 +44,14 @@ class ProfessoresController extends Controller
      */
     public function store(Request $request)
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_tcc) {
+            return response()->json([
+                "success"   => false,
+                "message"   => "Unauthorized."
+            ], 403);
+        }
+        
         $created = User::create([
             'name' => $request->input('name'),
             'email' => $request->input('email'),
@@ -68,6 +85,11 @@ class ProfessoresController extends Controller
      */
     public function show(User $professor)
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_users) {
+            return abort(404, 'Unauthorized');
+        }
+
         return view('professor.show',['professor' => $professor]);
     }
 
@@ -76,6 +98,11 @@ class ProfessoresController extends Controller
      */
     public function edit(User $professor)
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_users) {
+            return abort(404, 'Unauthorized');
+        }
+
         return view('professor.edit',['professor' => $professor]);
     }
 
@@ -84,6 +111,14 @@ class ProfessoresController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_tcc) {
+            return response()->json([
+                "success"   => false,
+                "message"   => "Unauthorized."
+            ], 403);
+        }
+
         $updated = User::where('id', $id)->update($request->except(['_token', '_method']));
 
         if($updated){
@@ -99,6 +134,14 @@ class ProfessoresController extends Controller
      */
     public function destroy(string $id)
     {
+        $group = auth()->user()->group()->first();
+        if(!$group->able_to_create_tcc) {
+            return response()->json([
+                "success"   => false,
+                "message"   => "Unauthorized."
+            ], 403);
+        }
+
         User::where('id', $id)->delete();
         return redirect()->route('professores-controller.index');
     }
